Refactor CodeGen to meet Python's PEP8 styleguide
